requirements.txt
asgiref==3.5.0
boto3==1.22.11
botocore==1.25.13
certifi==2021.10.8
cffi==1.15.0
commonmark==0.9.1
Django==4.0.3
django-cors-headers==3.11.0
django-rest==0.8.7
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
mysqlclient==2.1.0
oauthlib==3.2.0
PyJWT==2.3.0
PyMySQL==1.0.2
